guestagent: ebpfticker: impose a rate limit
Fix issue 4094

The `iptables` command invoked by the guestagent seems to cause
an extreme burst of `bind` events: issue 4094#issuecomment-3346519158

This commit introduces a rate limit so as to avoid consuming too much CPU
when a burst of events happen.

The `iptables` watcher is going to be soon removed (issue 4085),
but this rate limit seems still worth keeping after that.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>