Allow passing guestagent location to instance.StartWithPath
This is similar to passing in the full path to limactl to allow an embedding
application stored in a different location to launch a VM.

Simplified the internal calls by removing the limactl argument to instance.Start.

Also changed the order to look for external drivers first in LIMA_DRIVERS_PATH
before looking in libexec, allowing the user to override the drivers in the default
location. This also makes sure that the LIMA_DRIVERS_PATH drivers are found even
when libexec cannot be located.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>