Decouple diffdisk (misnomer) from basedisk
Fix issue 1706.

Now basedisk is immediately renamed/converted to diffdisk
(not a diff despite the name) by the VM driver,
except when basedisk is an ISO9660 image.

Decoupling diffdisk from basedisk will eliminate the overhead of differencing I/O
and save some disk space.

I cannot remember why I designed the disk to be split into basedisk and diffdisk.
Maybe it was to allow `limactl factory-reset` to retain the initial state,
although the command was apparently never implemented in that way.
Maybe it was to allow multiple instances to share the same basedisk,
although it was never implemented in that way, either.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>