hack/test-templates.sh: Loosen SSH over VSOCK detection tests
Detection of the SSH server on VSOCK may fail; however, a failing log indicates that controlling detection via the environment variable works as expected.

Change to allow failing detection of the SSH server on VSOCK, because it will fall back to connecting SSH over existing methods and work even if it fails.

Signed-off-by: Norio Nomura <norio.nomura@gmail.com>