feat: Add `limactl vz-vmnet-shared`
It shares `VmnetNetwork` serialization between VMs using SharedMode.
- `limactl vz-vmnet-shared --enable-mach-service`: register Mach service and launch
- `limactl vz-vmnet-shared --enable-mach-service=false`: unregister Mach service

When the `limactl` executable file is updated due to rebuilds, etc., the VM using the serialization data held by the Mach server before the update cannot be booted.
It is necessary to add a version check and restart the service as appropriate.
Also, it seems that it cannot be used with an external vz driver.

Signed-off-by: Norio Nomura <norio.nomura@gmail.com>