vz: Fix scrollbar issue with runtime swizzle
The Code-Hex/vz library wraps VZVirtualMachineView in an NSScrollView
with scrollbars hardcoded to YES (for their zoom feature). This causes
~16px of the VM display to be cut off by scrollbar gutters.

This patch uses Objective-C method swizzling to intercept NSScrollView's
setDocumentView: method and disable scrollbars when the document view
is a VZVirtualMachineView.

This is a self-contained fix that doesn't require forking the vz library.

Signed-off-by: stuffbucket <231133237+stuffbucket@users.noreply.github.com>